🌅🏂 Checkpoint ./tests/2_functional-tests.js:706117/284 ./routes/api.js:706117/6178

+++ routes/api.js
@@ -82,32 +82,39 @@ module.exports = function (app) {
   app.route('/api/threads/:board').delete((req,res)=>{
     //delete the thread document if password matches
     threadsCollection.findOne({_id:new ObjectId(req.body.thread_id), delete_password:req.body.delete_password}).then(thread=>{
-      //array of promises to delete thread, replies, and board reference
-      const promiseArr = [
-        //delete the replies
-        new Promise((resolve, reject)=>{
-          //get the _ids of each reply to be deleted
-          const replyIds = thread.replies.map(reply=>reply._id);
-          //delete each reply from the database
-          repliesCollection.deleteMany({_id:{$in:replyIds}}).then(()=>{
-            resolve();
-          }).catch(err=>reject(err));
-        }),
-        //update the board document
-        new Promise((resolve,reject)=>{
-          //remove the thread from the board document
-          boardsCollection.updateOne({name:req.params.board},{$pull:{threads:req.body.thread_id}}).then(()=>resolve()).catch(err=>reject(err));
-        }),
-        //remove the thread document
-        new Promise((resolve,reject)=>{
-          threadsCollection.deleteOne({_id:req.body.thread_id}).then(()=>resolve()).catch(err=>reject(err));
+      //if there was no thread found
+      if(!thread){
+        res.status(403).send('incorrect password');
+      }
+      else{
+        //array of promises to delete thread, replies, and board reference
+        const promiseArr = [
+          //delete the replies
+          new Promise((resolve, reject)=>{
+            //get the _ids of each reply to be deleted
+            const replyIds = thread.replies.map(reply=>reply._id);
+            //delete each reply from the database
+            repliesCollection.deleteMany({_id:{$in:replyIds}}).then(()=>{
+              resolve();
+            }).catch(err=>reject(err));
+          }),
+          //update the board document
+          new Promise((resolve,reject)=>{
+            //remove the thread from the board document
+            boardsCollection.updateOne({name:req.params.board},{$pull:{threads:req.body.thread_id}}).then(()=>resolve()).catch(err=>reject(err));
+          }),
+          //remove the thread document
+          new Promise((resolve,reject)=>{
+            threadsCollection.deleteOne({_id:req.body.thread_id}).then(()=>resolve()).catch(err=>reject(err));
+          })
+        ]
+        //when everything has been deleted
+        Promise.all(promiseArr).then(()=>{
+          res.status(200).send('success');
         })
-      ]
-      //when everything has been deleted
-      Promise.all(promiseArr).then(()=>{
-        res.status(200).send('success');
-      })
-    }).catch(err=>res.status(403).send('incorrect password'));
+      }
+      
+    }).catch(err=>res.status(400).json({error:err}));
   })
     
   app.route('/api/replies/:board').post((req,res)=>{
